I think the new event system has reached its completion for now
The new event system uses a custom EditorWindow accessed from Window/Event Manager.

From this window you have create, delete, and rename events. Then from code you can say EventIDList.*eventName* replacing *eventName* with a name of the event you added.

To create events you create a GameEventWrapper giving it the EventIDList.*eventName* as a parameter in its constructor. Then you call the GameEventWrapper's CreateEvent() function.

To invoke a created event use GameEventWrapper's Invoke function providing any desired parameters you wish to send. However, you cannot send multiple parameters of the same type. If you need to do that create a wrapper class that has them both as member variables.

To subscribe to an event use EventSystem.SubscribeToEvent() giving it the EventIDList.*eventName* of the event to subscribe to and the function you wish to be called by the event. The function must have a single parameter of type GameEventData. From this data you can use ReadValue<T>() to get the passed parameters of type T.

Unsubscribing is exactly the same as subscribing, but use EventSystem.UnsubscribeFromEvent().

If an event will be listened to all the time, it is recommended to subscribe to events in Unity's OnEnable function and to unsubscribe from events in Unity's OnDisable function.